+++ tests/MetaData/FacebookMetaDataTest.php
@@ -1,21 +1,34 @@
 <?php
 
+declare(strict_types=1);
+
 namespace SeoHelper\Tests\MetaData;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use SeoHelper\MetaData\BaseMetaData;
 use SeoHelper\MetaData\MetaData;
 
-class FacebookMetaDataTest extends PHPUnit_Framework_TestCase
+final class FacebookMetaDataTest extends TestCase
 {
-    public function testOgSiteName()
+    public function testOgSiteName(): void
     {
         $metaData = new MetaData();
         $this->assertTrue(is_array($metaData->get()));
-        $this->assertTrue(empty($metaData->get()));
-        
-        $this->assertFalse($metaData->getOgSiteName());
+        $this->assertEmpty($metaData->get());
+
+        $this->assertNull($metaData->getOgSiteName());
         $this->assertInstanceOf(BaseMetaData::class, $metaData->setOgSiteName('site-name'));
         $this->assertEquals(['site-name'], $metaData->getOgSiteName());
+    }
+
+    public function testFbAppId(): void
+    {
+        $metaData = new MetaData();
+        $this->assertTrue(is_array($metaData->get()));
+        $this->assertEmpty($metaData->get());
+
+        $this->assertNull($metaData->getFbAppId());
+        $this->assertInstanceOf(BaseMetaData::class, $metaData->setFbAppId('fb-app-id'));
+        $this->assertEquals(['fb-app-id'], $metaData->getFbAppId());
     }
 }